fix: Reduce default maxCVEsTotal from 10k to 5k to prevent gRPC message size limit
The gRPC transport has a 2MB message size limit, which was being exceeded
when SecurityResults contained large amounts of CVE data. Reducing the
default maxCVEsTotal from 10,000 to 5,000 CVEs per namespace should keep
the message size under the gRPC limit while still providing substantial
vulnerability data.

Error that prompted this change:
  rpc error: code = ResourceExhausted desc = trying to send message
  larger than max (2483073 vs. 2097152)

Note: This is in addition to the etcd 3MB limit. We now have two constraints:
- gRPC message size: ~2MB
- etcd object size: ~3MB

The gRPC limit is more restrictive, so we're optimizing for that.